+++ src/cmd/map.rs
@@ -1,7 +1,7 @@
 use std::convert::TryFrom;
 
 use crate::cmd::moonblade::{
-    run_moonblade_cmd, MoonbladeCmdArgs, MoonbladeErrorPolicy, MoonbladeMode,
+    run_moonblade_cmd, MoonbladeCmdArgs, LegacyMoonbladeErrorPolicy, MoonbladeMode,
 };
 use crate::config::Delimiter;
 use crate::util;
@@ -107,7 +107,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         no_headers: args.flag_no_headers,
         delimiter: args.flag_delimiter,
         parallelization,
-        error_policy: MoonbladeErrorPolicy::try_from(args.flag_errors)?,
+        error_policy: LegacyMoonbladeErrorPolicy::try_from(args.flag_errors)?,
         error_column_name: Some(args.flag_error_column),
         mode: MoonbladeMode::Map,
         ..Default::default()